Basic framework for a state machine. For now, make the robot stand still.
How it works:
assign controller.move to be any function that returns a Twist object.
every timestep, the controller will call controller.move() and send that
Twist object to the robot.

If you want to exit the script gracefully while the robot is moving,
the robot should stop automatically (we send a zero-vector Twist object
when there's a signal interrupt)